requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
python-dotenv
pandas
requests
werkzeug
